+++ packages/rs-drive-abci/Cargo.toml
@@ -42,6 +42,7 @@ dapi-grpc = { path = "../dapi-grpc", default-features = false, features = [
   "server",
   "platform",
 ] }
+platform-query-wire = { path = "../rs-platform-query-wire" }
 tracing-subscriber = { version = "0.3.22", default-features = false, features = [
   "env-filter",
   "ansi",
